+++ crates/cli/src/main.rs
@@ -8,7 +8,7 @@ async fn main() -> eyre::Result<()> {
     tracing_subscriber::fmt::fmt()
         .with_env_filter(
             EnvFilter::builder()
-                .with_default_directive(LevelFilter::INFO.into())
+                .with_default_directive(LevelFilter::WARN.into())
                 .from_env_lossy(),
         )
         .with_writer(std::io::stderr)