+++ crates/sdk/src/client/mod.rs
@@ -70,7 +70,7 @@ use gmsol_solana_utils::{
     bundle_builder::{BundleBuilder, BundleOptions, CreateBundleOptions},
     cluster::Cluster,
     program::Program,
-    transaction_builder::{Config, TransactionBuilder},
+    transaction_builder::{default_before_sign, Config, TransactionBuilder},
     utils::WithSlot,
 };
 use gmsol_utils::oracle::PriceProviderKind;
@@ -864,7 +864,8 @@ impl<C: Clone + Deref<Target = impl Signer>> Client<C> {
         );
         let status = view::view::<MarketStatus>(
             &self.store_program().rpc(),
-            &req.signed_transaction_with_options(true, None).await?,
+            &req.signed_transaction_with_options(true, None, default_before_sign)
+                .await?,
         )
         .await?;
         Ok(status)
@@ -882,7 +883,8 @@ impl<C: Clone + Deref<Target = impl Signer>> Client<C> {
         let req = self.get_market_token_price(store, market_token, prices, pnl_factor, maximize);
         let price = view::view::<u128>(
             &self.store_program().rpc(),
-            &req.signed_transaction_with_options(true, None).await?,
+            &req.signed_transaction_with_options(true, None, default_before_sign)
+                .await?,
         )
         .await?;
         Ok(price)